Rename models file.
