+++ functions/src/fulfillments/__tests__/roundFulfillment.test.ts
@@ -1,4 +1,4 @@
-import { ConversationData, ResponseType } from '../../models/models';
+import { ConversationData, ResponseType } from '../../models/conversation';
 import { chooseRound, roundHelpFulfillment } from '../roundFulfillment';
 
 describe('Select Round', () => {
