+++ functions/src/fulfillments/__tests__/welcomeFulfillment.test.ts
@@ -1,4 +1,4 @@
-import { ConversationData, ResponseType } from '../../models/models';
+import { ConversationData, ResponseType } from '../../models/conversation';
 
 import { invalidResponseFulfillment } from '../welcomeFulfillment';
