+++ functions/src/fulfillments/categoryFulfillment.ts
@@ -1,4 +1,8 @@
-import { ConversationData, Response, ResponseType } from '../models/models';
+import {
+  ConversationData,
+  Response,
+  ResponseType,
+} from '../models/conversation';
 import { OptionQuestion, Question } from '../models/questions';
 
 import { Category } from '../models/categories';
