+++ functions/src/fulfillments/trueFalseFulfillment.ts
@@ -3,7 +3,7 @@ import {
   Response,
   ResponseType,
   Unknown,
-} from '../models/models';
+} from '../models/conversation';
 import { OptionQuestion, Question, QuestionType } from '../models/questions';
 import {
   buildSSMLAndCombineAudioResponses,
